decouple validate and ocr
