+++ ocr/__init__.py
@@ -2,7 +2,5 @@
 from .ocr import extract_text
 from .ocr import extract_string
 from .ocr import text_to_dataframe
-from .ocr import format_df
 from .ocr import create_name_candidates
-from .ocr import look_for_name
